+++ CHANGELOG.md
@@ -1,12 +1,18 @@
 # unreleased
+## Fixes
 * Bump version of scryfall-client to v0.8.0
 
+## Features
+* Add automatic paging of search results
+
 # 0.2.0
+## Features
 * Add ability to/from import/export with TappedOut.net files
 * Add tag support
 * Move card preview to be inline in card list
 
 # 0.1.0
+## Features
 * Add import and export buttons in settings modal
 
 # 0.0.0